main/base/viewsets.py
from rest_framework import generics
from rest_framework.viewsets import ViewSetMixin


class GenericViewSet(ViewSetMixin, generics.GenericAPIView):
    """
    The GenericViewSet class does not provide any actions by default,
    but does include the base set of generic view behavior, such as
    the `get_object` and `get_queryset` methods.
    """

    def initial(self, request, *args, **kwargs):
        """
        Runs anything that needs to occur prior to calling the method handler.
        """

        super(GenericViewSet, self).initial(request, *args, **kwargs)
