requirements.txt
appdirs==1.4.4
asgiref==3.3.4
autopep8==1.5.7
backcall==0.2.0
cfgv==3.2.0
decorator==5.0.8
distlib==0.3.1
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
drf-nested-routers==0.93.3
filelock==3.0.12
identify==2.2.4
ipdb==0.13.7
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
nodeenv==1.6.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pre-commit==2.12.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.9.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
virtualenv==20.4.6
wcwidth==0.2.5
